requirements.txt
# ============================================
# Project Sol: 依存パッケージ一覧 (Ver 11.0)
# ============================================
# 
# 【推奨インストール方法】
# エラーが発生する場合は、段階的にインストールしてください:
#   1. pip install -r requirements-core.txt
#   2. pip install -r requirements-optional.txt
#
# 【一括インストール（エラーが出る場合あり）】
#   pip install -r requirements.txt
#
# または、仮想環境を使用する場合:
#   python3 -m venv venv
#   source venv/bin/activate  # Linux/Mac
#   venv\Scripts\activate     # Windows
#   pip install -r requirements-core.txt
#   pip install -r requirements-optional.txt
# ============================================

# ============================================
# コア依存パッケージ（必須）
# ============================================

# 環境変数管理
python-dotenv>=1.0.0

# データ検証（Pydantic）
pydantic>=2.0.0

# ============================================
# AI/LLM API（必須 - いずれか1つ以上）
# ============================================

# Google Gemini API（推奨 - 新しいアーキテクチャ）
google-generativeai>=0.8.0

# OpenAI API（レガシー - 後方互換性のため）
openai>=1.0.0

# ============================================
# LangChain エコシステム（RAG機能用）
# ============================================

# LangChain コア
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# ============================================
# ベクトルストア（RAG機能用 - オプション）
# ============================================

# ChromaDB（ローカルベクトルストア）
chromadb>=0.4.0

# ============================================
# 埋め込みモデル（ローカル - オプション）
# ============================================

# Sentence Transformers（ローカル埋め込み）
sentence-transformers>=2.2.0

# ============================================
# 日本語テキスト処理（オプション）
# ============================================

# BudouX（日本語分かち書き）
budoux>=0.7.0

# ============================================
# リトライロジック（推奨）
# ============================================

# Tenacity（指数バックオフリトライ）
tenacity>=8.2.0

# ============================================
# データベース（デプロイ用）
# ============================================

# Supabase（データベースデプロイ）
supabase>=2.3.0

# ============================================
# Docker操作（オプション）
# ============================================

# Docker Python SDK（コンテナビルド・テスト用）
docker>=6.0.0

# ============================================
# HTTPリクエスト（オプション）
# ============================================

# Requests（HTTPリクエスト - コンテナテスト用）
requests>=2.31.0

# ============================================
# その他のユーティリティ
# ============================================

# JSON処理（標準ライブラリだが、明示的に記載）
# json - 標準ライブラリ（インストール不要）

# ============================================
# 開発・テスト用（オプション）
# ============================================

# pytest>=7.0.0  # テストフレームワーク（必要に応じて）
# black>=23.0.0  # コードフォーマッター（必要に応じて）
# mypy>=1.0.0    # 型チェッカー（必要に応じて）

# ============================================
# ローカルLLM（オプション - フォールバック用）
# ============================================
# 
# インターネット接続がない場合やAPI制限に達した場合の
# フォールバックとして使用できます。
# 
# インストール方法:
#   pip install llama-cpp-python
# 
# モデルダウンロード:
#   https://huggingface.co/google/gemma-2-2b-jpn-it
# 
# llama-cpp-python>=0.2.0  # コメントアウト（必要に応じて有効化）

